pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "negclone"
version = "0.1.0"
description = "Extract grain/color fingerprints from film scans and generate Darktable/Lightroom presets"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "Desert Paul" },
]

dependencies = [
    "typer>=0.15.0",
    "rich>=13.9.0",
    "Pillow>=11.0.0",
    "numpy>=2.1.0",
    "scikit-image>=0.24.0",
    "flickrapi>=2.4.0",
    "httpx>=0.28.0",
    "pydantic>=2.10.0",
]

[project.scripts]
negclone = "negclone.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "respx>=0.22.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E", "F", "W", "I", "N", "UP", "B", "A", "SIM", "TCH",
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=negclone --cov-report=term-missing"
